+++ src/middlewares/ForceSsl.ts
@@ -6,7 +6,7 @@ import { isLocalIpAddress } from "@lorenstuff/universal-utilities";
 
 import { FritterContext } from "../classes/FritterContext.js";
 
-import { MiddlewareFunction } from "../types/MiddlewareFunction.js";
+import { FritterMiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
 
 //
 // Middleware
@@ -22,7 +22,7 @@ export type CreateOptions =
 export type CreateResult =
 {
 	allowInsecureLocalIpAddresses: boolean;
-	execute: MiddlewareFunction<MiddlewareFritterContext>;
+	execute: FritterMiddlewareFunction<MiddlewareFritterContext>;
 };
 
 export function create(options?: CreateOptions): CreateResult