+++ src/types/FritterHttpMethod.ts
@@ -1,0 +1,10 @@
+//
+// Types
+//
+
+/**
+ * An HTTP method.
+ *
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods
+ */
+export type FritterHttpMethod = "GET" | "HEAD" | "POST" | "PUT" | "DELETE" | "CONNECT" | "OPTIONS" | "TRACE" | "PATCH";