+++ src/types/PossibleJsonData.ts
@@ -1,6 +1,0 @@
-//
-// Type
-//
-
-/** @deprecated */
-export type PossibleJsonData = string | number | boolean | null | { [key: string]: PossibleJsonData } | PossibleJsonData[];